Move cert fixture to filesystem test folder (#814)
